Move code out of Program
Move as much of the entry-point logic out of `Program` as possible.
